src/requirements.txt
requests
pandas
boto3
openai
python-telegram-bot
instaloader
pydantic>=2.0